requirements.txt
django
python-dotenv
djangorestframework
drf-spectacular
drf-spectacular-sidecar
djangorestframework-simplejwt
django-cors-headers
requests
Pillow
PyJWT
django-ratelimit
python-dateutil